requirements.txt
python-dotenv
SQLAlchemy>=1.4
aiomysql
pymysql
alpaca_trade_api
ratelimit
tenacity
aiohttp
feedparser
requests
ta
scikit-learn
xgboost
lightgbm
pandas
numpy
matplotlib
scipy
statsmodels
shap
aio_pika
transformers
prometheus_client
fredapi
APScheduler
rich
openai
psycopg2-binary
ta-lib
